+++ corda/workflows/src/main/kotlin/nl/tno/federated/corda/flows/EventFlows.kt
@@ -1,4 +1,4 @@
-package nl.tno.federated.flows
+package nl.tno.federated.corda.flows
 
 import co.paralleluniverse.fibers.Suspendable
 import net.corda.core.contracts.requireThat
@@ -10,14 +10,16 @@ import net.corda.core.transactions.TransactionBuilder
 import net.corda.core.utilities.ProgressTracker
 import net.corda.core.utilities.ProgressTracker.Step
 import nl.tno.federated.contracts.EventContract
+import nl.tno.federated.corda.services.ishare.ISHARECordaService
+import nl.tno.federated.ishare.config.ISHAREConfig
 import nl.tno.federated.states.EventState
 import org.slf4j.LoggerFactory
 
 @InitiatingFlow
 @StartableByRPC
 class NewEventFlow(
-    val destinations: Collection<CordaX500Name>,
-    val event: String
+    private val destinations: Collection<CordaX500Name>,
+    private val event: String
 ) : FlowLogic<SignedTransaction>() {
 
     private val log = LoggerFactory.getLogger(NewEventFlow::class.java)
@@ -45,6 +47,8 @@ class NewEventFlow(
             GATHERING_SIGS,
             FINALISING_TRANSACTION
         )
+
+        val ishareConfig = ISHAREConfig.loadProperties("ishare.properties")
     }
 
     override val progressTracker = tracker()
@@ -84,6 +88,22 @@ class NewEventFlow(
 
         // Stage 3.
         progressTracker.currentStep = SIGNING_TRANSACTION
+
+        /*
+          When a Event is send he same  event is send to all parties
+          This needs to change, because if ISHARE is used every party hands out his own accesstoken, which is stored in the event.
+          If we could store the accesstoken in the session maybe it will persist.
+        */
+        if (ishareConfig.enabled) {
+            // create an eventstate for every party , since they all have different accesstokens
+
+            counterParties.forEach {
+                //  create new session to get the tokens ?
+                val tokenResponse = subFlow(ISHARETokenFlow(it))
+                newEventState.accessTokens[it] = tokenResponse.access_token
+            }
+        }
+
         // Sign the transaction.
         val partSignedTx = serviceHub.signInitialTransaction(txBuilder)
 
@@ -145,9 +165,25 @@ class NewEventResponder(val counterpartySession: FlowSession) : FlowLogic<Signed
     @Suspendable
     override fun call(): SignedTransaction {
         progressTracker.currentStep = VERIFYING_STRING_INTEGRITY
+
+
         val signTransactionFlow = object : SignTransactionFlow(counterpartySession) {
             override fun checkTransaction(stx: SignedTransaction) = requireThat {
                 val outputState = stx.tx.outputStates.single() as EventState
+
+                // Check accesstoken if required and available in the eventState
+                if (serviceHub.cordaService(ISHARECordaService::class.java).ishareEnabled()) {
+                    require(outputState.accessTokens.contains(serviceHub.myInfo.legalIdentities.first())) {
+                        "ISHARE accessToken not found while it is required"
+                    }
+                    outputState.accessTokens[serviceHub.myInfo.legalIdentities.first()]?.let {
+                        require(serviceHub.cordaService(ISHARECordaService::class.java).checkAccessToken(it).first) {
+                            "ISHARE AccessToken is invalid."
+                        }
+                        // TODO optional check if insert of event is allowed by the iSHARE AR
+                    }
+                }
+
                 require(
                     graphdb().insertEvent(
                         outputState.fullEvent,