+++ corda/workflows/src/main/kotlin/nl/tno/federated/corda/flows/GraphDBFlows.kt
@@ -1,10 +1,12 @@
-package nl.tno.federated.flows
+package nl.tno.federated.corda.flows
 
 import co.paralleluniverse.fibers.Suspendable
+import net.corda.core.flows.FlowExternalOperation
 import net.corda.core.flows.FlowLogic
 import net.corda.core.flows.InitiatingFlow
 import net.corda.core.flows.StartableByRPC
-import nl.tno.federated.services.CordaGraphDBService
+import nl.tno.federated.corda.services.graphdb.GraphDBCordaService
+import nl.tno.federated.corda.services.ishare.ISHARECordaService
 import nl.tno.federated.states.Event
 
 @InitiatingFlow
@@ -15,6 +17,7 @@ class QueryGraphDBbyIdFlow(
 
     @Suspendable
     override fun call(): String {
+        // TODO https://docs.r3.com/en/platform/corda/4.9/community/api-flows.html#calling-external-systems-inside-of-flows
         return graphdb().queryEventById(id)
     }
 }
@@ -27,19 +30,8 @@ class GeneralSPARQLqueryFlow(
 
     @Suspendable
     override fun call(): String {
+        // TODO https://docs.r3.com/en/platform/corda/4.9/community/api-flows.html#calling-external-systems-inside-of-flows
         return graphdb().generalSPARQLquery(query)
-    }
-}
-
-@InitiatingFlow
-@StartableByRPC
-class ParseRDFFlow(
-    val rdfEvent: String
-) : FlowLogic<List<Event>>() {
-
-    @Suspendable
-    override fun call(): List<Event> {
-        return graphdb().parseRDFToEvents(rdfEvent)
     }
 }
 
@@ -52,8 +44,9 @@ class InsertRDFFlow(
 
     @Suspendable
     override fun call(): Boolean {
+        // TODO https://docs.r3.com/en/platform/corda/4.9/community/api-flows.html#calling-external-systems-inside-of-flows
         return graphdb().insertEvent(rdfEvent, privateRepo)
     }
 }
 
-fun FlowLogic<*>.graphdb() = serviceHub.cordaService(CordaGraphDBService::class.java)
+fun FlowLogic<*>.graphdb() = serviceHub.cordaService(GraphDBCordaService::class.java)